chore(graphql): update variable names
