+++ packages/graphql/src/liveProvider/index.ts
@@ -2,15 +2,15 @@ import { LiveProvider } from "@pankod/refine-core";
 import { Client } from "graphql-ws";
 
 import {
-    genereteUseListSubscription,
-    genereteUseManySubscription,
-    genereteUseOneSubscription,
+    generateUseListSubscription,
+    generateUseManySubscription,
+    generateUseOneSubscription,
 } from "../utilities";
 
 const subscriptions = {
-    useList: genereteUseListSubscription,
-    useOne: genereteUseOneSubscription,
-    useMany: genereteUseManySubscription,
+    useList: generateUseListSubscription,
+    useOne: generateUseOneSubscription,
+    useMany: generateUseManySubscription,
 };
 
 export const liveProvider = (client: Client): LiveProvider => {
